requirements.txt
asttokens==2.0.5
blinker==1.8.2
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
comm==0.2.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
debugpy==1.6.7
decorator==5.1.1
easyocr==1.7.0
EasyProcess==1.1
entrypoint2==1.1
executing==0.8.3
filelock==3.13.1
Flask==3.0.3
fonttools==4.52.4
fsspec==2024.3.1
gmpy2==2.1.2
grpcio==1.64.1
grpcio-tools==1.64.1
idna==3.7
imagecodecs==2023.1.23
imageio==2.33.1
intel-openmp==2021.4.0
ipykernel==6.28.0
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.0
jupyter_core==5.7.2
kiwisolver==1.4.5
lapx==0.5.9
lazy_loader==0.4
marko==2.1.2
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mkl==2021.4.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
mss==9.0.1
munkres==1.1.4
mutagen==1.47.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0
opencv-python-headless==4.10.0
packaging==24.0
pandas==2.2.2
parso==0.8.3
patsy==0.5.6
pillow==10.4.0
pip==24.0
platformdirs==3.10.0
ply==3.11
prompt-toolkit==3.0.43
protobuf==5.27.1
psutil==5.9.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyclipper==1.3.0.post3
pycryptodomex==3.20.0
Pygments==2.15.1
pyparsing==3.1.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyscreenshot==3.1
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.9.0
python-slugify==8.0.4
pytz==2024.1
pywin32==305.1
PyYAML==6.0.1
pyzmq==24.0.1
requests==2.32.2
scikit-image==0.23.2
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
setuptools==69.5.1
shapely==2.0.5
simplejpeg==1.7.2
sip==6.7.12
six==1.16.0
stack-data==0.2.0
statsmodels==0.14.2
sympy==1.12
tbb==2021.12.0
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2023.4.12
toml==0.10.2
tomli==2.0.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.24
urllib3==2.2.1
vidgear==0.3.2
wcwidth==0.2.5
webcolors==24.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
win-inet-pton==1.1.0
yt-dlp==2024.5.27
